requirements.txt
azure-eventhub==1.3.*
schedule
paho-mqtt